poc/a_communication/poc_1_web_socket/requirements.txt
# WebSocket PoC Dependencies
fastapi==0.115.6
uvicorn[standard]==0.34.0
websockets==14.1
PyJWT==2.10.1